requirements.txt
customtkinter>=5.2.0
tkinterdnd2>=0.4.0
faster-whisper>=1.0.0
google-genai>=1.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
python-vlc>=3.0.0
